User notice for parser migration; add preference for article views
This patch adds `$wgParserMigrationEnableParsoidArticlePages` to enable
Parsoid read views by default for *article pages*, complementing the
existing `$wgParserMigrationEnableParsoidDiscussionTools` which enables
Parsoid read views by default for *talk page*.

It also adds a configurable user notification the first time a particular
user visits a page rendered with Parsoid, either by opting in or when
the wiki default provides for it.  The notification will only show up
after `$wgParserMigrationNoticeDays` once dismissed, or when
`$wgParserMigrationNoticeVersion` is incremented (indicating a
substantial change to the notice which merits redisplay).

This patch also updates the README.md to describe more fully the
present operation of this extension.

Bug: T355567
Bug: T355568
Change-Id: If762b61780f1ec59e50b6a2bd2badb8d3aea0311