Create "report visual bug" dialog
This is disabled by default to ease deployment.  To enable, set
$wgParserMigrationEnableReportVisualBug to true.

The default is to add reports to the wikipage [[Project:Parsoid/Feedback]]
on the local wiki, which can be customized via the
[[MediaWiki:Parsermigration-reportbug-feedback-title]] system message.

However, $wgParserMigrationFeedbackTitle and
$wgParserMigrationFeedbackAPIURL can be set to enable centralized
reporting to a single wiki page, for example on mediawiki.org.

The backend code is based on mw.Feedback.  Notable features:

* We add a change tag for the edit, `parsermigration-visual-bug`, so that
  reports created by this tool can be pulled out of a recent changes feed.

* We suppress captchas on the feedback page configured for this tool, and
  suppress change tags for users without the 'applychangetags' permission,
  maximizing the likelihood that reports can be recorded successfully.

* Links are provided to the feedback page after submission so the user
  can view or further edit their report.

* Supports MinervaNeue/MobileFrontend as well as desktop skins.

Bug: T365371
Change-Id: If7669acca51430392e57b77b6886f9b3921c70cf
(cherry picked from commit 74192a9ca480deb3817d0139371ca2b112372194)